moved feature point graph values to graph localizer
