+++ localization/graph_localizer/include/graph_localizer/projection_factor_adder.h
@@ -19,10 +19,10 @@
 #ifndef GRAPH_LOCALIZER_PROJECTION_FACTOR_ADDER_H_
 #define GRAPH_LOCALIZER_PROJECTION_FACTOR_ADDER_H_
 
+#include <graph_localizer/feature_point_graph_values.h>
 #include <graph_localizer/feature_tracker.h>
 #include <graph_localizer/projection_factor_adder_params.h>
 #include <graph_optimizer/factor_adder.h>
-#include <graph_optimizer/feature_point_graph_values.h>
 #include <localization_measurements/feature_points_measurement.h>
 
 #include <gtsam/geometry/triangulation.h>
@@ -38,14 +38,14 @@ class ProjectionFactorAdder : public graph_optimizer::FactorAdder<localization_m
  public:
   ProjectionFactorAdder(const ProjectionFactorAdderParams& params,
                         std::shared_ptr<const FeatureTracker> feature_tracker,
-                        std::shared_ptr<const graph_optimizer::FeaturePointGraphValues> feature_point_graph_values);
+                        std::shared_ptr<const FeaturePointGraphValues> feature_point_graph_values);
 
   std::vector<graph_optimizer::FactorsToAdd> AddFactors(
     const localization_measurements::FeaturePointsMeasurement& feature_points_measurement) final;
 
  private:
   std::shared_ptr<const FeatureTracker> feature_tracker_;
-  std::shared_ptr<const graph_optimizer::FeaturePointGraphValues> feature_point_graph_values_;
+  std::shared_ptr<const FeaturePointGraphValues> feature_point_graph_values_;
 };
 }  // namespace graph_localizer
