+++ localization/graph_localizer/include/graph_localizer/projection_graph_action_completer.h
@@ -21,7 +21,7 @@
 
 #include <graph_localizer/combined_nav_state_graph_values.h>
 #include <graph_localizer/projection_factor_adder_params.h>
-#include <graph_optimizer/feature_point_graph_values.h>
+#include <graph_localizer/feature_point_graph_values.h>
 #include <graph_optimizer/graph_action_completer.h>
 
 #include <gtsam/geometry/triangulation.h>
@@ -31,7 +31,7 @@ class ProjectionGraphActionCompleter : public graph_optimizer::GraphActionComple
  public:
   ProjectionGraphActionCompleter(const ProjectionFactorAdderParams& params,
                                  std::shared_ptr<const CombinedNavStateGraphValues> graph_values,
-                                 std::shared_ptr<graph_optimizer::FeaturePointGraphValues> feature_point_graph_values);
+                                 std::shared_ptr<FeaturePointGraphValues> feature_point_graph_values);
 
   bool DoAction(graph_optimizer::FactorsToAdd& factors_to_add, gtsam::NonlinearFactorGraph& graph_factors) final;
 
@@ -42,7 +42,7 @@ class ProjectionGraphActionCompleter : public graph_optimizer::GraphActionComple
 
   ProjectionFactorAdderParams params_;
   std::shared_ptr<const CombinedNavStateGraphValues> graph_values_;
-  std::shared_ptr<graph_optimizer::FeaturePointGraphValues> feature_point_graph_values_;
+  std::shared_ptr<FeaturePointGraphValues> feature_point_graph_values_;
   gtsam::TriangulationParameters projection_triangulation_params_;
 };
 }  // namespace graph_localizer
